Move the save method from JsonLDHandlerTest to MetacatHandlerTest since JsonLDHander doesn't have save now.
